refactor: move PlayerItinerary to player package
- Updated AndroidManifest to reflect new package structure.
- Renamed PlayerItinerary class and updated its imports.
- Adjusted references in InfoItinerary for consistency.
